improve guest time sync with host
Time in VirtualBox guest VM's can fall out of sync when the host machine is
suspended and resumed.  You can see this by suspending a host for a few
minutes, wake it up, and compare host and guest clocks.

It seems VirtualBox will not take action until 20min of drift is seen.  This
change will reduce the threshold to 10s.  Details here:
http://stackoverflow.com/questions/19490652/how-to-sync-time-on-host-wake-up-within-virtualbox#19492466

The default value can't be seen but after setting, you can do this to see it:

    $ VBoxManage guestproperty get a1.dcos "/VirtualBox/GuestAdd/VBoxService/--timesync-set-threshold"
    Value: 10000